refactored feedItem to be more broken up for more refined syling options
